internal/state/state: Add Leader client helper to State
Adds a new function for getting a client connected to the leader of the
dqlite cluster.

Some additional changes:

Fixed a comment for state.Cluster, informing that it doesn't return a
client for the entire cluster, just every node except ourselves.

Additionally passes our server cert as the client cert.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>