cluster/stmt: Split registered statements by caller package
If a microcluster project indirectly imports microcluster from another
project (i.e. microcloud using microceph, which both use microcluster),
then the map of registered statements may end up polluted with entries
from that project.

To avoid this, a call to microcluster.Start will record the name of the
caller package (whether from snap build, GOPATH, or go module), and
record it in the daemon.

As well, on start, when calls to `RegisterStmt` are executed, those
packages will be recorded, and statements will be registered by their
package.

Then, when preparing statements, only statements belonging to the daemon
package, or to 'microcluster' (as that is ubiquitous) will be prepared.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>