internal/rest/resources: Add endpoint for registering local cluster records on join
Previously, local records weren't updated until the next heartbeat,
except on the system that handled the join request. This adds a new
method PUT to /cluster which will forward a request to every system,
instructing it to add the newly added node to its store.

As a result, the node forwarding a join request to a leader will no
longer also implicitly trust whoever sent the request, as the token
won't have been validated, and the endpoint does not have
authentication. Join requests now have the RestrictNotification
AccessHandler set so a request sent with the cluster notification flag
will go through authentication.

Lastly, as none of the methods on this endpoint should be called while
the node is uninitialied, `AllowedBeforeInit` is set to false.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>