internal/trust: Clean up StartAPI
This moves some of the logic in StartAPI out into other helpers to make
it easier to read, and prevent duplication of those actions.

This includes detecting when a cluster is finished comparing its schema
version to other members, and obtaining clients for the whole cluster to
send a notification.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>