internal/db/update: Split internal and external schema versions
Reworks the schema package with a new schema update that splits the
`version` column of the `schemas` table into `internal_version` and
`external_version` so that we can properly handle updates to both.

Because this schema update affects the actual schema update logic, it
also needs to be called in 2 places during the schema version check
logic so that both pre-existing and new clusters behave the same.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>