microcluster: Block until cancelled in the daemon package
By handling all sources of cancellation in the daemon package itself,
and using `signal.NotifyContext`, it becomes possible to simplify the
blocking loop considerably.

There are 3 ways to cancel the daemon:

 - Receive a signal to terminate, cancelling `ctx` from microcluster.App
   and resulting in us calling d.stop, and sending to the
   `d.shutdownDoneCh` to notify the blocking loop that we can return.

 - The parent ctx of `ctx` from microcluster.App cancels for an unknown
   reason. `ctx` will also cancel so the situation is the same as above.

 - `POST /cluster/control/shutdown` from the unix socket will shut down
   the daemon via the REST API. In this case we wait to send the to the
   `d.shutdownDoneCh` until we can be sure the daemon will receive the
   response with the correct error code.

As `d.stop` is a `sync.Once`, we can directly call it from both the API
and daemon directly and it will block until the first one returns, so
all we have to check before calling `d.stop` is whether the parent
context was cancelled.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>