internal/daemon: Send separate schema/api upgrade notifications
If we have both API and schema upgrades to consider, the first node to
consider all other nodes ready to upgrade will apply the schema upgrade,
and then notify the other nodes that it has done so. This frees up all
other nodes to record their expected API extension count, and the last
node to detect that this has been performed on all nodes will notify the
cluster that the API extension count is verified.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>