internal/rest/types: Save trusted system in token response
After a system uses its join token to reach out to a cluster member, the
dqlite leader establishes trust with this node and sends back
information that the joiner needs to establish communication with the
rest of the cluster.

This adds that existing cluster member's info to this set of information
so that the joiner can tell which node already has the joiner's record
in its truststore.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>