rest/access: Add external access package
Moves the authentication handlers out into their own function that is
exported and can be used to facilitate finer control of authentication
per-endpoint.

Additionally allows passing in a map of x509.Certificate to authenticate
against a set other than the default truststore, if desired. If used
externally in a project specific AccessHandler, the request context will
still report that the request is untrusted by the truststore, unless the
certificate used to authenticate the request also exists in the
truststore. In either case, the request will still go through.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>