internal/db/update: Also set legacy_alter_table when running internal updates
This will disable the behaviour of sqlite where ALTER TABLE RENAME
will only rename the initial occurrence of the table name in
CREATE TABLE, CREATE INDEX, and CREATE TRIGGER statements.

This will make it so existing views or tables with WHERE or CHECK
clauses will not fail during compilation, and can be fixed in a
subsequent external schema update.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>