internal/db/update: Add schema update for applying the first API extensions
Adds a schema update that applies API extensions automatically once we
are certain that all systems support API extensions.

Because schema updates are only committed once all systems have been
updated, it is safe to apply the first set of API extensions right away.

If for some reason each system has a different set of API extensions,
the synchronization step that verifies all systems have the same set of
API extensions will continue to run after applying this schema update,
before the cluster is notified that it is ready to start.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>