internal/rest/resources: Split resetClusterMember into its own helper
We can use resetClusterMember to clear the state if there is an error
during joins as well. This requires adding the option to skip all
errors with the `force` query parameter.

In order to get around the synchronization of the re-execing of the
daemon, a function is returned that can be run in a go-routine, which
locks the cluster disable mutex and re-execs the daemon.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>