cluster: Create a temporary table to hold the real schema update count
So that we don't break existing online cluster members, we can't just
update the schema from underneath them.

Instead, we can store schema updates in a temporary table
`internal_cluster_members_new`, and reference this table during the
schema update process, instead of the real `internal_cluster_members`
table.

Then, when we actually run the `updateFromV1` on the final cluster
member to receive the update, replace the real
`internal_cluster_members` table with the one that has been keeping
track of the updates.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>