internal/daemon: Initialize extension servers based on core configuration
On init, the flow of server initialization is now slightly different.

Prior to init:
- If `ListenPort` is set on the daemon, we start all core servers
  listening on this port. If any extension servers have PreInit set to
  true, they will also be started with this port.

- If any extension servers have a defined address, they will be started
  here as well.

After init:
- All core endpoints will be restarted using the cluster cert & daemon
  address. Any extension servers lacking an address & certificate will
  be restarted with the core configuration.

- Any extension servers with their own defined address will be started.

In general, this allows for skipping some extension servers if the
`init` command is not populated with an address for those servers, as
long as those servers have their own defined cert. Otherwise, they will
be lumped in with the core servers.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>