internal/daemon: Attempt to re-use control.socket if present
If the daemon crashed or stopped uncleanly due to an error, then the
control socket file might still be present in the state directory. This
means the daemon will enter a restart loop until the control socket is
manually removed.

Instead, we can just attempt to re-connect to the control socket. If
it's already in use then that will error out the daemon anyway.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>