internal/db: Pre-emptively prepare API extensions before waiting for schema update
When a node updates its schema version, it may learn that it needs to
wait because another node is behind. The node should also update its
expected set of API extensions at this time, so that once a node detects
all nodes are ready for the schema update, we don't have to wait longer
for API extensions to settle.

That said, we may not have API extension support so we have to manually
check for it before attempting the update.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>