internal/recover: Update database/cluster.yaml directly
When starting go-dqlite with out-of-date cluster.yaml addresses, we end
up exceeding a context deadline waiting on db.Ready(). I'm assuming
go-dqlite is pulling the listen addrs from cluster.yaml instead of the
dqlite raft log.

We only need to write the file once since it is then included in the
recovery tarball and copied to all the other members.

Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>