internal/rest/resources: Improve join logic to handle unreachable nodes
Fixes an issue where the join process would fail if a node in the
join_address list is unreachable.

When a token is created, it includes the IP addresses of all nodes in the cluster.
If a node is removed between the token creation and joining, the join process previously
failed because it tried to connect to the non-existent node.

Now, the client discovery logic will continue attempting to connect to other nodes in the
join_address list instead of failing early. This ensures that the join process can succeed
as long as at least one node is reachable.

Signed-off-by: Benjamin Schimke <benjamin.schimke@canonical.com>