cluster: Parse go module version out of project name
This helper finds microcluster among the cached build files to register
internal statements. Since microcluster's module path has been appended
with v2, this function now resolves to `v2` as the project name instead
of `microcluster`, so this commit properly checks if the path is of the
form v[number] and if so, looks one directory back instead.

This should fix external projects failing to register internal
microcluster prepared statements.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>