internal/rest/resources: Fix cluster member re-exec on error
During cluster join if an error is returned from any of the operations (e.g. hooks), the revert statements
ensures that the cluster member is reset accordingly.
However it's important to not call the re-exec operation twice as it will interrupt all ongoing connections
from the first re-exec attempt.
To not step on the foot of the first re-exec operation, it only gets executed in case we are actually joining a cluster.
In all other cases we don't perform the manual re-exec but rely on the cluster member deletion endpoint to
trigger the final re-exec as part of the deletion of the cluster member.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit c88eaade10949b6a5d1c3f350ad40d947ebdf3ac)