internal/rest: Support proxying websocket connections
When enabling ProxyTarget on the EndpointAction, Microcluster forwards the request to another cluster member in case the target query paramter is specified.
In case the request contains a websocket upgrade header, the proxy is required to support proxying the websocket connection.
This can be done by opening up another websocket to the target side and proxying both connections back and forth.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>